requirements.txt
amqp==5.2.0
anyio==4.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
boto3==1.34.44
botocore==1.34.44
CacheControl==0.14.0
celery==5.3.6
certifi==2024.2.2
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.0.2
django-leaflet==0.29.0
djangorestframework==3.14.0
geographiclib==2.0
h11==0.14.0
httptools==0.6.1
idna==3.6
iso3166==2.1.1
jmespath==1.0.1
kombu==5.3.5
mapbox==0.18.1
msgpack==1.0.7
pillow==10.2.0
polyline==2.0.2
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
s3transfer==0.10.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.0.7
uvicorn==0.27.1
vine==5.1.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
